port vmmap to aglib
