+++ pwndbg/aglib/__init__.py
@@ -22,6 +22,7 @@ def load_aglib():
     import pwndbg.aglib.file
     import pwndbg.aglib.heap
     import pwndbg.aglib.kernel
+    import pwndbg.aglib.kernel.vmmap
     import pwndbg.aglib.memory
     import pwndbg.aglib.nearpc
     import pwndbg.aglib.next
@@ -35,6 +36,7 @@ def load_aglib():
     import pwndbg.aglib.symbol
     import pwndbg.aglib.typeinfo
     import pwndbg.aglib.vmmap
+    import pwndbg.aglib.vmmap_custom
 
     # This is necessary so that mypy understands the actual type of the regs module
     regs_: regs_mod.module = regs_mod
