+++ pwndbg/dbg/gdb/__init__.py
@@ -383,11 +383,23 @@ class GDBProcess(pwndbg.dbg_mod.Process):
     @override
     def vmmap(self) -> pwndbg.dbg_mod.MemoryMap:
         import pwndbg.aglib.qemu
-        import pwndbg.gdblib.vmmap
+        from pwndbg.aglib.kernel.vmmap import kernel_vmmap
+        from pwndbg.aglib.vmmap_custom import get_custom_pages
+        from pwndbg.gdblib.vmmap import get_known_maps
 
-        pages = pwndbg.gdblib.vmmap.get()
         qemu = pwndbg.aglib.qemu.is_old_qemu_user()
+        proc_maps = get_known_maps()
+        # The `proc_maps` is usually a tuple of Page objects but it can also be:
+        #   None    - when /proc/$tid/maps does not exist/is not available
+        #   tuple() - when the process has no maps yet which happens only during its very early init
+        #             (usually when we attach to a process)
+        if proc_maps is not None:
+            return GDBMemoryMap(qemu, proc_maps)
 
+        pages: List[pwndbg.lib.memory.Page] = []
+        pages.extend(kernel_vmmap())
+        pages.extend(get_custom_pages())
+        pages.sort()
         return GDBMemoryMap(qemu, pages)
 
     @override
